requirements.txt
langchain==1.2.13
langchain-community==0.4.1
langchain-google-genai==4.2.1
chromadb==1.5.5
pypdf==6.9.2
google-generativeai==0.8.6
streamlit==1.55.0
python-dotenv==1.2.2
